+++ src/main/kotlin/com/atlassian/performance/tools/awsinfrastructure/api/network/access/ForSecurityGroupAccessRequester.kt
@@ -1,0 +1,12 @@
+package com.atlassian.performance.tools.awsinfrastructure.api.network.access
+
+import com.amazonaws.services.ec2.model.SecurityGroup
+
+class ForSecurityGroupAccessRequester(
+    private val securityGroupProvider: () -> SecurityGroup
+) : AccessRequester {
+    override fun requestAccess(
+        accessProvider: AccessProvider
+    ) = accessProvider
+        .provideAccess(securityGroupProvider())
+}