+++ src/main/kotlin/com/atlassian/performance/tools/awsinfrastructure/api/network/access/NoAccessProvider.kt
@@ -1,0 +1,13 @@
+package com.atlassian.performance.tools.awsinfrastructure.api.network.access
+
+import com.amazonaws.services.ec2.model.SecurityGroup
+
+class NoAccessProvider : AccessProvider {
+    override fun provideAccess(
+        cidr: String
+    ) = true
+
+    override fun provideAccess(
+        securityGroup: SecurityGroup
+    ) = true
+}