+++ src/main/kotlin/com/atlassian/performance/tools/awsinfrastructure/api/network/access/NoAccessRequester.kt
@@ -1,0 +1,7 @@
+package com.atlassian.performance.tools.awsinfrastructure.api.network.access
+
+class NoAccessRequester : AccessRequester {
+    override fun requestAccess(
+        accessProvider: AccessProvider
+    ) = true
+}