+++ src/test/kotlin/com/atlassian/performance/tools/awsinfrastructure/TemplateBuilderTest.kt
@@ -6,30 +6,6 @@ import org.junit.Assert.assertThat
 import org.junit.Test
 
 class TemplateBuilderTest {
-
-    @Test
-    fun shouldAddSecurityGroupIngress() {
-        val port = 9997
-        val baseTemplate = "single-node.yaml"
-        val originalTemplate = TemplateBuilder(baseTemplate).build()
-
-        val changedTemplate = TemplateBuilder(baseTemplate)
-            .let {
-                it.addSecurityGroupIngress(
-                    newResourceKey = "TestIngress",
-                    groupId = it.getResourceReference("JiraNodeSecurityGroup"),
-                    ports = (port..port)
-                )
-            }
-            .build()
-
-        val expectedOutputs = listOf("FromPort: $port", "ToPort: $port")
-        for (expectedOutput in expectedOutputs) {
-            assertThat(originalTemplate, not(containsString(expectedOutput)))
-            assertThat(changedTemplate, containsString(expectedOutput))
-        }
-    }
-
     @Test
     fun shouldSetNodesCount() {
         val nodesCount = 4
@@ -40,6 +16,4 @@ class TemplateBuilderTest {
             assertThat(changedTemplate, containsString("jira$nodeNumber"))
         }
     }
-
-
 }